I implemented enough functionality so that dirty files would get saved on deactivation. Here is how I did it.
First I changed the mixin from top-level-window<%> to drscheme:unit:frame<%>. The reason is that a top level window is either a frame% or a dialog%:

file:///C:/Program%20Files/PLT/doc/mred/mred-Z-H-277.html#node_tag_Temp_279

Clicking through the hierarchy of drscheme:unit:frame% on:

http://download.plt-scheme.org/doc/tools/tools-Z-H-171.html#node_sec_3.29

we find that:

drscheme:unit:frame%->drscheme:frame:mixin->drscheme:frame:<%>->frame:editor<%>->frame:standard-menus<%>->frame:basic<%>->(class->interface frame%)->frame%.

We need top-level-window so we can override on-activate. Next, we need to be able to ask the frame for its tabs. We are mising into drscheme:unit:frame<%>:

http://download.plt-scheme.org/doc/tools/tools-Z-H-158.html#node_idx_738

Now we can call get-tabs to get a list of drscheme:unit:tab<%>.

http://download.plt-scheme.org/doc/tools/tools-Z-H-158.html#node_idx_738

On each tab we call get-defs to get a text% object.

http://download.plt-scheme.org/doc/tools/tools-Z-H-209.html#node_tag_Temp_210

text% implements editor<%>, and provides everything you would expect like get-filename, is-modified?, and save-file.

The way that this function works is that when the frame is deactivated, it iterates over all of the tabs in the frame, and if file exists for the tab (one does not have to exist, the user may have never saved one) and the file has been modified, the the file gets saved with the pre-existing filename and format, and any errors that occur are displayed.