Refactoring each-tab.
Updated each-tab so that it takes two thunks, a predicate and an action, both take an argument of an editor. If the predicate is true, the action is called. Both handlers were updated to pass two functions to each-tab. Two helper functions were created, file-loaded? and file-modified?, both take an editor, and do as names imply.