Last night I realized that no one will ever use this tool if after reverting, the cursor position gets reset. Today I read the editor<%> documentation and that made me think I need to use a position object. There are functions get-start-position and set-position. I knew that DrScheme already does revert, so I looked at the source for how they do it in there. See:
collects\framework\private\frame.ss line 1147.

Observation: the calls to the editor get-filename use the boxed boolean and check the result, this seems redundant to me, and I didn't do it that way. The function always checks if the editor is a text% before using it. In the case of this tool, it is assumed that each editor is a text%. Perhaps there is something I don't know, the question is "Could a tab ever not be a text%?". This is something for future investigation.

frame.ss revert and mine are similar enough, they both reload the file. frame's goes the extra step and resets the cursor position, and also does a performance tweak by stopping and starting edit refreshs to the screen. I implemented this tool's revert in a similar manner.

Should revert be implemented twice, in DrScheme and this tool? I don't know.